+++ src/bika/lims/browser/analyses/view.py
@@ -102,6 +102,7 @@ class AnalysesView(ListingView):
         self.dmk = context.bika_setup.getResultsDecimalMark()
         self.scinot = context.bika_setup.getScientificNotationResults()
         self.categories = []
+        self.expand_all_categories = True
 
         # each editable item needs it's own allow_edit
         # which is a list of field names.
@@ -239,6 +240,8 @@ class AnalysesView(ListingView):
         super(AnalysesView, self).update()
         self.load_analysis_categories()
         self.append_partition_filters()
+        if self.analysis_categories_enabled():
+            self.show_categories = True
 
     def before_render(self):
         """Before render hook
@@ -268,6 +271,16 @@ class AnalysesView(ListingView):
         """Check if analysis remarks are enabled
         """
         return self.context.bika_setup.getEnableAnalysisRemarks()
+
+    @viewcache.memoize
+    def analysis_categories_enabled(self):
+        """Check if analyses should be grouped by category
+        """
+        # setting applies only for samples
+        if not IAnalysisRequest.providedBy(self.context):
+            return False
+        setup = api.get_senaite_setup()
+        return setup.getCategorizeSampleAnalyses()
 
     @viewcache.memoize
     def has_permission(self, permission, obj=None):
@@ -768,6 +781,14 @@ class AnalysesView(ListingView):
 
         return items
 
+    def render_unit(self, unit, css_class=None):
+        """Render HTML element for unit
+        """
+        if css_class is None:
+            css_class = "unit d-inline-block py-2 small text-secondary"
+        return "<span class='{css_class}'>{unit}</span>".format(
+            unit=unit, css_class=css_class)
+
     def _folder_item_category(self, analysis_brain, item):
         """Sets the category to the item passed in
 
@@ -858,6 +879,11 @@ class AnalysesView(ListingView):
         item["CaptureDate"] = capture_date_str
         item["result_captured"] = capture_date_str
 
+        # Add the unit after the result
+        unit = item.get("Unit")
+        if unit:
+            item["after"]["Result"] = self.render_unit(unit)
+
         # Get the analysis object
         obj = self.get_object(analysis_brain)
 
@@ -887,7 +913,7 @@ class AnalysesView(ListingView):
 
             else:
                 item["result_type"] = "numeric"
-                item["help"]["result"] = _(
+                item["help"]["Result"] = _(
                     "Enter the result either in decimal or scientific "
                     "notation, e.g. 0.00005 or 1e-5, 10000 or 1e5")
 
@@ -953,10 +979,17 @@ class AnalysesView(ListingView):
                 continue
 
             interim_value = interim_field.get("value", "")
+            interim_unit = interim_field.get("unit", "")
             interim_formatted = formatDecimalMark(interim_value, self.dmk)
             interim_field["formatted_value"] = interim_formatted
             item[interim_keyword] = interim_field
             item["class"][interim_keyword] = "interim"
+
+            # render the unit after the interim field
+            if interim_unit:
+                formatted_interim_unit = format_supsub(interim_unit)
+                item["after"][interim_keyword] = self.render_unit(
+                    formatted_interim_unit)
 
             # Note: As soon as we have a separate content type for field
             #       analysis, we can solely rely on the field permission
@@ -1506,11 +1539,11 @@ class AnalysesView(ListingView):
         """
         # Always return true if the analysis has a method assigned
         obj = self.get_object(analysis)
-        method = obj.getMethod()
+        method = obj.getRawMethod()
         if method:
             return True
 
-        methods = obj.getAllowedMethods()
+        methods = obj.getRawAllowedMethods()
         return len(methods) > 0
 
     def is_instrument_required(self, analysis):
@@ -1522,13 +1555,13 @@ class AnalysesView(ListingView):
         # If method selection list is required, the instrument selection too
         if self.is_method_required(analysis):
             return True
-        
+
         # Always return true if the analysis has an instrument assigned
-        if self.get_instrument(analysis):
+        analysis = self.get_object(analysis)
+        if analysis.getRawInstrument():
             return True
 
-        obj = self.get_object(analysis)
-        instruments = obj.getAllowedInstruments()
+        instruments = analysis.getRawAllowedInstruments()
         # There is no need to check for the instruments of the method assigned
         # to # the analysis (if any), because the instruments rendered in the
         # selection list are always a subset of the allowed instruments when