+++ src/bika/lims/browser/widgets/referenceresultswidget.py
@@ -291,6 +291,8 @@ class ReferenceResultsWidget(TypesWidget):
         table.before_render()
 
         if allow_edit is False:
+            # This is a hack to notify read-only mode to the view
+            table.allow_edit = allow_edit
             return table.contents_table_view()
 
         return table.ajax_contents_table()