+++ src/senaite/core/upgrade/v02_03_000.py
@@ -27,8 +27,12 @@ from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.core.catalog import SETUP_CATALOG
 from senaite.core.catalog.report_catalog import ReportCatalog
 from senaite.core.config import PROJECTNAME as product
+from senaite.core.setuphandlers import CATALOG_MAPPINGS
 from senaite.core.setuphandlers import _run_import_step
 from senaite.core.setuphandlers import add_senaite_setup
+from senaite.core.setuphandlers import setup_auditlog_catalog_mappings
+from senaite.core.setuphandlers import setup_catalog_mappings
+from senaite.core.setuphandlers import setup_catalogs_order
 from senaite.core.setuphandlers import setup_core_catalogs
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
@@ -47,7 +51,7 @@ METADATA_TO_REMOVE = [
 @upgradestep(product, version)
 def upgrade(tool):
     portal = tool.aq_inner.aq_parent
-    setup = portal.portal_setup  # noqa
+    setup = portal.portal_setup
     ut = UpgradeUtils(portal)
     ver_from = ut.getInstalledVersion(product)
 
@@ -70,6 +74,16 @@ def upgrade(tool):
     setup.runImportStepFromProfile(profile, "plone.app.registry")
     setup.runImportStepFromProfile(profile, "controlpanel")
 
+    # Ensure the catalog mappings for Analyses and Samples is correct
+    # https://github.com/senaite/senaite.core/pull/2130
+    setup_catalog_mappings(portal, catalog_mappings=CATALOG_MAPPINGS)
+
+    # remap auditlog catalog
+    setup_auditlog_catalog_mappings(portal)
+
+    # ensure the catalogs assigned to types are sorted correctly
+    setup_catalogs_order(portal)
+
     # Add new setup folder to portal
     add_senaite_setup(portal)
 
@@ -82,6 +96,7 @@ def upgrade(tool):
     move_arreports_to_report_catalog(portal)
     migrate_analysis_services_fields(portal)
     migrate_analyses_fields(portal)
+    reindex_laboratory(portal)
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
@@ -244,7 +259,7 @@ def fix_unassigned_samples(portal):
         cat.catalog_object(obj, obj_url, idxs=indexes, update_metadata=1)
 
         # Flush the object from memory
-        obj._p_deactivate()  # noqa
+        obj._p_deactivate()
 
     logger.info("Fix unassigned samples ...")
 
@@ -281,7 +296,7 @@ def move_arreports_to_report_catalog(portal):
         obj.reindexObject()
 
         # Flush the object from memory
-        obj._p_deactivate()  # noqa
+        obj._p_deactivate()
 
     logger.info("Move ARReports to SENAITE Report Catalog [DONE]")
 
@@ -308,7 +323,7 @@ def migrate_analysis_services_fields(portal):
         migrate_ldl_field_to_string(obj)
 
         # Flush the object from memory
-        obj._p_deactivate()  # noqa
+        obj._p_deactivate()
 
     logger.info("Migrate Analysis Services [DONE]")
 
@@ -338,7 +353,7 @@ def migrate_analyses_fields(portal):
         migrate_ldl_field_to_string(obj)
 
         # Flush the object from memory
-        obj._p_deactivate()  # noqa
+        obj._p_deactivate()
 
     logger.info("Migrate Analyses Fields [DONE]")
 
@@ -411,3 +426,12 @@ def fixed_point_value_to_string(value, precision):
     str_value = template % (sign, front, fra)
     # strip off trailing zeros and possible dot
     return str_value.rstrip("0").rstrip(".")
+
+
+def reindex_laboratory(portal):
+    """Forces the reindex of laboratory content type
+    """
+    logger.info("Reindexing laboratory content type ...")
+    setup = api.get_setup()
+    setup.laboratory.reindexObject()
+    logger.info("Reindexing laboratory content type [DONE]")