+++ CHANGES.rst
@@ -28,11 +28,37 @@ Changelog
 
 2.1.0 (2022-01-05)
 ------------------
-
-
-2.3.0 (unreleased)
+2.4.0 (unreleased)
 ------------------
 
+
+
+2.3.0 (2022-10-03)
+------------------
+
+- #2153 Improve default body text for email publication to allow translations
+- #2151 Added `api.copy_object` function for both DX and AT types
+- #2150 Improve the creation process of AT content types
+- #2150 Added `api.edit` function for both DX and AT types
+- #2149 Performance: prioritize raw getter for AllowedMethods field
+- #2148 Performance: prioritize raw getter for AllowedInstruments field
+- #2147 Remove stale function workflow.getReviewHistory
+- #2146 Fix "No object found for UID: <laboratory_uid>" in report preview
+- #2145 Crop page navigation for DX reference widget
+- #2143 Fix Traceback when using readonly decorator for objects w/o __name__
+- #2140 Allow to enable/disable analysis categories for samples
+- #2137 Dynamic Workflow Menu
+- #2139 Fix LabClerk cannot create partitions from received samples
+- #2130 Catalog mapping for Samples and Analyses
+- #2131 Allow to edit the analysis service sort keys in the services listing
+- #2133 Filter Contact and CCContact by Client on first click
+- #2134 Render unit after result and interim fields
+- #2132 Add a default date for batch creation 
+- #2129 Fix Traceback when invalidating a Sample with Remarks
+- #2128 Fix referenceresults widget view mode
+- #2127 Fix instrument expiry date display in listing view
+- #2123 Add Sample Form: Save and Copy Action
+- #2119 Fix linked client contact user can not see existing samples
 - #2118 Customized Quickinstaller Configlet
 - #2117 Customized User/Groups Preferences in Site Configuration
 - #2115 Fix insufficient privileges for instrument QC viewlet